requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
networkx==3.1
osmnx==1.6.0
numpy==1.24.3
shapely==2.0.1
google-api-python-client==2.108.0
google-auth==2.50.0
python-dotenv==1.0.0
scikit-learn==1.3.0
